Moved Backdrop and motion.div from ModalCard.jsx into Modal.jsx. Renamed ModalCard.jsx to EventModal.jsx.
