+++ client/src/features/modal/Modal.jsx
@@ -1,7 +1,9 @@
 import { useContext } from "react";
 import { AnimatePresence } from "framer-motion";
-import ModalCard from "./ModalCard";
+import EventModal from "./EventModal";
 import { Context } from "contexts/Context";
+import Backdrop from "./Backdrop";
+import { motion } from "framer-motion";
 
 const Modal = () => {
   const [context, setContext] = useContext(Context)
@@ -11,6 +13,27 @@ const Modal = () => {
     setContext({ ...context })
   }
 
+  const dropIn = {
+    hidden: {
+      y: "-100vh",
+      opacity: 0,
+    },
+    visible: {
+      y: "0",
+      opacity: 1,
+      transition: {
+        duration: 0.1,
+        type: "spring",
+        damping: 25,
+        stiffness: 500,
+      },
+    },
+    exit: {
+      y: "-100vh",
+      opacity: 0,
+    },
+  };
+
   return (
     <div>
       <AnimatePresence
@@ -18,7 +41,20 @@ const Modal = () => {
         exitBeforeEnter={true}
         onExitComplete={() => null}
       >
-        {context.modalOpen && <ModalCard modalOpen={context.modalOpen} handleClose={toggleModal} />}
+        {context.modalOpen &&
+          <Backdrop onClick={toggleModal}>
+            <motion.div
+              className="modal"
+              onClick={e => e.stopPropagation()}
+              variants={dropIn}
+              initial="hidden"
+              animate="visible"
+              exit="exit"
+            >
+              <EventModal modalOpen={context.modalOpen} handleClose={toggleModal} />
+            </motion.div>
+          </Backdrop>
+        }
       </AnimatePresence>
     </div>
   );
